Sending JSON data from client to webmethod
We can now send information TO the webmethod. It comes in as a parameter
to the called webmethod and you can do whatever with it.  For now it's
pretty basic, but take a look at the returnTwoWay method in Service.cs
and its corresponding client side javascript-ajax-function in
default.aspx (sendJSONTest()).

basically sendJSONTest calls returnTwoWay webmethod but also sends it
json data. the returnTwoWay webmethod gets that data as a parameter.