Snap to the grid
So essentally we can take our graphs but them on these "widgets" and
move them around the place. They allow you to snap to the grid and it
expands them as you move them around. NO resizing avaliable yet will be
working on it and it is not integrated with tabs things soon to come.
BTW its awesome